Dynamically test configure role based on host/group vars
It would have been easier to modify the host_vars to include the
hostname.
But that would have been a dependency between the filename and the
hostname inside that file that should match the file name.

This moves the responsibility to the molecule test without any additions
to the code, to keep the test hostname mapping up to date